Решение ДЗ 9 Рефакторинг
