+++ TinkoffApp/AppDelegate.swift
@@ -16,9 +16,8 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         // Override point for customization after application launch.
         FirebaseApp.configure()
-        UserIDManager.loadId()
         // Читаем все сохраненные данные
-        CoreDataManager.readAllData()
+        RootAssembly.serviceAssembly.coreData.readAllData()
         return true
     }
