+++ TinkoffApp/CoreLayer/Models/CoreData/Chat.xcdatamodeld/Chat.xcdatamodel/contents
