--- TinkoffApp/Models/ConversationCellModel.swift
@@ -1,14 +1,0 @@
-//
-//  ConversationCellModel.swift
-//  TinkoffApp
-//
-//  Created by Михаил on 29.09.2020.
-//  Copyright © 2020 Tinkoff. All rights reserved.
-//
-
-import Foundation
-struct ConversationCellModel {
-    let name: String
-    let message: String
-    let date: Date
-}
