+++ TinkoffApp/PresentationLayer/Modules/Protocols/ThemesPickerDelegate.swift
@@ -1,0 +1,12 @@
+//
+//  ThemesPickerDelegate.swift
+//  TinkoffApp
+//
+//  Created by Михаил on 07.10.2020.
+//  Copyright © 2020 Tinkoff. All rights reserved.
+//
+
+import Foundation
+protocol ThemesPickerDelegate: NSObject {
+    func applyTheme()
+}
